@@ -70,6 +70,6 @@ class ReadingsController < ApplicationController
     end
 
     def reading_params
-      params.require(:reading).permit(:equipment_id, :humedad_origen1, :humedad_origen2, :humedad_origen3, :temperatura_origen, :humedad1, :humedad2, :humedad3, :temperatura)
+      params.require(:reading).permit([:equipment_id, :humedad_origen1, :humedad_origen2, :humedad_origen3, :temperatura_origen, :humedad1, :humedad2, :humedad3, :temperatura]).to_h
     end
 end
