@@ -13947,124 +13947,145 @@
 
 
 
-AstronomicObject.create(name: 'M1', equivalence: 'NGC1952', type_object: 'Sn', bmag:8.4, ra: '5h 34.5m', dec: '+22° 01′', constellation: 'Tau', colloquial_name: 'Crab Nebula');
-AstronomicObject.create(name: 'M2', equivalence: 'NGC7089', type_object: 'Gc', bmag:6.5, ra: '21h 33.5m', dec: '+00° 49′', constellation: 'Aqr', colloquial_name: '');
-AstronomicObject.create(name: 'M3', equivalence: 'NGC5272', type_object: 'Gc', bmag:6.2, ra: '13h 42.2m', dec: '+28° 23′', constellation: 'CVn', colloquial_name: '');
-AstronomicObject.create(name: 'M4', equivalence: 'NGC6121', type_object: 'Gc', bmag:5.6, ra: '16h 23.6m', dec: '-26° 32′', constellation: 'Sco', colloquial_name: '');
-AstronomicObject.create(name: 'M5', equivalence: 'NGC5904', type_object: 'Gc', bmag:5.6, ra: '15h 18.6m', dec: '+02° 05′', constellation: 'Ser', colloquial_name: '');
-AstronomicObject.create(name: 'M6', equivalence: 'NGC6405', type_object: 'Oc', bmag:4.2, ra: '17h 40.1m', dec: '-32° 13′', constellation: 'Sco', colloquial_name: 'Butterfly Cluster');
-AstronomicObject.create(name: 'M7', equivalence: 'NGC6475', type_object: 'Oc', bmag:3.3, ra: '17h 53.9m', dec: '-34° 49′', constellation: 'Sco', colloquial_name: 'Ptolemys Cluster');
-AstronomicObject.create(name: 'M8', equivalence: 'NGC6523', type_object: 'Di', bmag:6, ra: '18h 03.8m', dec: '-24° 23′', constellation: 'Sgr', colloquial_name: 'Lagoon Nebula');
-AstronomicObject.create(name: 'M9', equivalence: 'NGC6333', type_object: 'Gc', bmag:7.7, ra: '17h 19.2m', dec: '-18° 31′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M10', equivalence: 'NGC6254', type_object: 'Gc', bmag:6.6, ra: '16h 57.1m', dec: '-04° 06′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M11', equivalence: 'NGC6705', type_object: 'Oc', bmag:6.3, ra: '18h 51.1m', dec: '-06° 16′', constellation: 'Sct', colloquial_name: 'Wild Duck Cluster');
-AstronomicObject.create(name: 'M12', equivalence: 'NGC6218', type_object: 'Gc', bmag:6.7, ra: '16h 47.2m', dec: '-01° 57′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M13', equivalence: 'NGC6205', type_object: 'Gc', bmag:5.8, ra: '16h 41.7m', dec: '+36° 28′', constellation: 'Her', colloquial_name: 'Great Hercules Globular');
-AstronomicObject.create(name: 'M14', equivalence: 'NGC6402', type_object: 'Gc', bmag:7.6, ra: '17h 37.6m', dec: '-03° 15′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M15', equivalence: 'NGC7078', type_object: 'Gc', bmag:6.2, ra: '21h 30m', dec: '+12° 10′', constellation: 'Peg', colloquial_name: 'Great Pegasus Globular');
-AstronomicObject.create(name: 'M16', equivalence: 'NGC6611', type_object: 'Oc', bmag:6.4, ra: '18h 18.8m', dec: '-13° 47′', constellation: 'Ser', colloquial_name: 'Eagle Nebula');
-AstronomicObject.create(name: 'M17', equivalence: 'NGC6618', type_object: 'Di', bmag:7, ra: '18h 20.8m', dec: '-16° 11′', constellation: 'Sgr', colloquial_name: 'Omega Nebula');
-AstronomicObject.create(name: 'M18', equivalence: 'NGC6613', type_object: 'Oc', bmag:7.5, ra: '18h 19.9m', dec: '-17° 08′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M19', equivalence: 'NGC6273', type_object: 'Gc', bmag:6.8, ra: '17h 02.6m', dec: '-26° 16′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M20', equivalence: 'NGC6514', type_object: 'Di', bmag:9, ra: '18h 02.6m', dec: '-23° 02′', constellation: 'Sgr', colloquial_name: 'Trifid Nebula');
-AstronomicObject.create(name: 'M21', equivalence: 'NGC6531', type_object: 'Oc', bmag:6.5, ra: '18h 04.6m', dec: '-22° 30′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M22', equivalence: 'NGC6656', type_object: 'Gc', bmag:5.1, ra: '18h 36.4m', dec: '-23° 54′', constellation: 'Sgr', colloquial_name: 'Sagittarius Cluster');
-AstronomicObject.create(name: 'M23', equivalence: 'NGC6494', type_object: 'Oc', bmag:6.9, ra: '17h 56.8m', dec: '-19° 01′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M24', equivalence: '', type_object: 'MW', bmag:4.6, ra: '18h 16.9m', dec: '-18° 30′', constellation: 'Sgr', colloquial_name: 'Sagittarius Star Cloud');
-AstronomicObject.create(name: 'M25', equivalence: 'IC4725', type_object: 'Oc', bmag:6.5, ra: '18h 31.6m', dec: '-19° 15′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M26', equivalence: 'NGC6694', type_object: 'Oc', bmag:8, ra: '18h 45.2m', dec: '-09° 24′', constellation: 'Sct', colloquial_name: '');
-AstronomicObject.create(name: 'M27', equivalence: 'NGC6853', type_object: 'Pl', bmag:7.4, ra: '19h 59.6m', dec: '+22° 43′', constellation: 'Vul', colloquial_name: 'Dumbbell Nebula');
-AstronomicObject.create(name: 'M28', equivalence: 'NGC6626', type_object: 'Gc', bmag:6.8, ra: '18h 24.5m', dec: '-24° 52′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M29', equivalence: 'NGC6913', type_object: 'Oc', bmag:7.1, ra: '20h 23.9m', dec: '+38° 32′', constellation: 'Cyg', colloquial_name: '');
-AstronomicObject.create(name: 'M30', equivalence: 'NGC7099', type_object: 'Gc', bmag:7.2, ra: '21h 40.4m', dec: '-23° 11′', constellation: 'Cap', colloquial_name: '');
-AstronomicObject.create(name: 'M31', equivalence: 'NGC224', type_object: 'Sp', bmag:3.4, ra: '0h 41.8m', dec: '+41° 16′', constellation: 'And', colloquial_name: 'Andromeda Galaxy');
-AstronomicObject.create(name: 'M32', equivalence: 'NGC221', type_object: 'El', bmag:8.1, ra: '0h 42.8m', dec: '+40° 52′', constellation: 'And', colloquial_name: '');
-AstronomicObject.create(name: 'M33', equivalence: 'NGC598', type_object: 'Sp', bmag:5.7, ra: '1h 33.9m', dec: '+30° 39′', constellation: 'Tri', colloquial_name: 'Triangulum Galaxy');
-AstronomicObject.create(name: 'M34', equivalence: 'NGC1039', type_object: 'Oc', bmag:5.5, ra: '2h 42m', dec: '+42° 47′', constellation: 'Per', colloquial_name: '');
-AstronomicObject.create(name: 'M35', equivalence: 'NGC2168', type_object: 'Oc', bmag:5.3, ra: '6h 08.9m', dec: '+24° 20′', constellation: 'Gem', colloquial_name: '');
-AstronomicObject.create(name: 'M36', equivalence: 'NGC1960', type_object: 'Oc', bmag:6.3, ra: '5h 36.1m', dec: '+34° 08′', constellation: 'Aur', colloquial_name: '');
-AstronomicObject.create(name: 'M37', equivalence: 'NGC2099', type_object: 'Oc', bmag:6.2, ra: '5h 52.4m', dec: '+32° 33′', constellation: 'Aur', colloquial_name: '');
-AstronomicObject.create(name: 'M38', equivalence: 'NGC1912', type_object: 'Oc', bmag:7.4, ra: '5h 28.7m', dec: '+35° 50′', constellation: 'Aur', colloquial_name: '');
-AstronomicObject.create(name: 'M39', equivalence: 'NGC7092', type_object: 'Oc', bmag:4.6, ra: '21h 32.2m', dec: '+48° 26′', constellation: 'Cyg', colloquial_name: '');
-AstronomicObject.create(name: 'M40', equivalence: '', type_object: 'Ds', bmag:8.4, ra: '12h 22.4m', dec: '+58° 05′', constellation: 'UMa', colloquial_name: 'Winnecke 4');
-AstronomicObject.create(name: 'M41', equivalence: 'NGC2287', type_object: 'Oc', bmag:4.6, ra: '6h 47m', dec: '-20° 44′', constellation: 'CMa', colloquial_name: '');
-AstronomicObject.create(name: 'M42', equivalence: 'NGC1976', type_object: 'Di', bmag:4, ra: '5h 35.4m', dec: '-05° 27′', constellation: 'Ori', colloquial_name: 'Great Nebula in Orion');
-AstronomicObject.create(name: 'M43', equivalence: 'NGC1982', type_object: 'Di', bmag:9, ra: '5h 35.6m', dec: '-05° 16′', constellation: 'Ori', colloquial_name: 'De Mairans Nebula');
-AstronomicObject.create(name: 'M44', equivalence: 'NGC2632', type_object: 'Oc', bmag:3.7, ra: '8h 40.1m', dec: '+19° 59′', constellation: 'Cnc', colloquial_name: 'Beehive Cluster');
-AstronomicObject.create(name: 'M45', equivalence: '', type_object: 'Oc', bmag:1.6, ra: '3h 47m', dec: '+24° 07′', constellation: 'Tau', colloquial_name: 'Pleiades');
-AstronomicObject.create(name: 'M46', equivalence: 'NGC2437', type_object: 'Oc', bmag:6, ra: '7h 41.8m', dec: '-14° 49′', constellation: 'Pup', colloquial_name: '');
-AstronomicObject.create(name: 'M47', equivalence: 'NGC2422', type_object: 'Oc', bmag:5.2, ra: '7h 36.6m', dec: '-14° 30′', constellation: 'Pup', colloquial_name: '');
-AstronomicObject.create(name: 'M48', equivalence: 'NGC2548', type_object: 'Oc', bmag:5.5, ra: '8h 13.8m', dec: '-05° 48′', constellation: 'Hya', colloquial_name: '');
-AstronomicObject.create(name: 'M49', equivalence: 'NGC4472', type_object: 'El', bmag:8.4, ra: '12h 29.8m', dec: '+08° 00′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M50', equivalence: 'NGC2323', type_object: 'Oc', bmag:6.3, ra: '7h 03.2m', dec: '-08° 20′', constellation: 'Mon', colloquial_name: '');
-AstronomicObject.create(name: 'M51', equivalence: 'NGC5194', type_object: 'Sp', bmag:8.4, ra: '13h 30m', dec: '+47° 11′', constellation: 'CVn', colloquial_name: 'Whirlpool Galaxy');
-AstronomicObject.create(name: 'M52', equivalence: 'NGC7654', type_object: 'Oc', bmag:7.3, ra: '23h 24.2m', dec: '+61° 35′', constellation: 'Cas', colloquial_name: '');
-AstronomicObject.create(name: 'M53', equivalence: 'NGC5024', type_object: 'Gc', bmag:7.6, ra: '13h 12.9m', dec: '+18° 10′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M54', equivalence: 'NGC6715', type_object: 'Gc', bmag:7.6, ra: '18h 55.1m', dec: '-30° 29′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M55', equivalence: 'NGC6809', type_object: 'Gc', bmag:6.3, ra: '19h 40m', dec: '-30° 58′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M56', equivalence: 'NGC6779', type_object: 'Gc', bmag:8.3, ra: '19h 16.6m', dec: '+30° 11′', constellation: 'Lyr', colloquial_name: '');
-AstronomicObject.create(name: 'M57', equivalence: 'NGC6720', type_object: 'Pl', bmag:8.8, ra: '18h 53.6m', dec: '+33° 02′', constellation: 'Lyr', colloquial_name: 'Ring Nebula');
-AstronomicObject.create(name: 'M58', equivalence: 'NGC4579', type_object: 'Ba', bmag:9.7, ra: '12h 37.7m', dec: '+11° 49′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M59', equivalence: 'NGC4621', type_object: 'El', bmag:9.6, ra: '12h 42m', dec: '+11° 39′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M60', equivalence: 'NGC4649', type_object: 'El', bmag:8.8, ra: '12h 43.7m', dec: '+11° 33′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M61', equivalence: 'NGC4303', type_object: 'Sp', bmag:9.7, ra: '12h 21.9m', dec: '+04° 28′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M62', equivalence: 'NGC6266', type_object: 'Gc', bmag:6.5, ra: '17h 01.2m', dec: '-30° 07′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M63', equivalence: 'NGC5055', type_object: 'Sp', bmag:8.6, ra: '13h 15.8m', dec: '+42° 02′', constellation: 'CVn', colloquial_name: 'Sunflower Galaxy');
-AstronomicObject.create(name: 'M64', equivalence: 'NGC4826', type_object: 'Sp', bmag:8.5, ra: '12h 56.7m', dec: '+21° 41′', constellation: 'Com', colloquial_name: 'Black Eye Galaxy');
-AstronomicObject.create(name: 'M65', equivalence: 'NGC3623', type_object: 'Sp', bmag:9.3, ra: '11h 18.9m', dec: '+13° 05′', constellation: 'Leo', colloquial_name: '');
-AstronomicObject.create(name: 'M66', equivalence: 'NGC3627', type_object: 'Sp', bmag:8.9, ra: '11h 20.2m', dec: '+12° 59′', constellation: 'Leo', colloquial_name: '');
-AstronomicObject.create(name: 'M67', equivalence: 'NGC2682', type_object: 'Oc', bmag:6.1, ra: '8h 50.4m', dec: '+11° 49′', constellation: 'Cnc', colloquial_name: '');
-AstronomicObject.create(name: 'M68', equivalence: 'NGC4590', type_object: 'Gc', bmag:7.8, ra: '12h 39.5m', dec: '-26° 45′', constellation: 'Hya', colloquial_name: '');
-AstronomicObject.create(name: 'M69', equivalence: 'NGC6637', type_object: 'Gc', bmag:7.6, ra: '18h 31.4m', dec: '-32° 21′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M70', equivalence: 'NGC6681', type_object: 'Gc', bmag:7.9, ra: '18h 43.2m', dec: '-32° 18′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M71', equivalence: 'NGC6838', type_object: 'Gc', bmag:8.2, ra: '19h 53.8m', dec: '+18° 47′', constellation: 'Sge', colloquial_name: '');
-AstronomicObject.create(name: 'M72', equivalence: 'NGC6981', type_object: 'Gc', bmag:9.3, ra: '20h 53.5m', dec: '-12° 32′', constellation: 'Aqr', colloquial_name: '');
-AstronomicObject.create(name: 'M73', equivalence: 'NGC6994', type_object: 'As', bmag:9, ra: '20h 59m', dec: '-12° 38′', constellation: 'Aqr', colloquial_name: '');
-AstronomicObject.create(name: 'M74', equivalence: 'NGC628', type_object: 'Sp', bmag:9.4, ra: '1h 36.7m', dec: '+15° 47′', constellation: 'Psc', colloquial_name: '');
-AstronomicObject.create(name: 'M75', equivalence: 'NGC6864', type_object: 'Gc', bmag:8.5, ra: '20h 06.1m', dec: '-21° 55′', constellation: 'Sgr', colloquial_name: '');
-AstronomicObject.create(name: 'M76', equivalence: 'NGC650', type_object: 'Pl', bmag:10.1, ra: '1h 42.4m', dec: '+51° 34′', constellation: 'Per', colloquial_name: 'Little Dumbbell Nebula');
-AstronomicObject.create(name: 'M77', equivalence: 'NGC1068', type_object: 'Sp', bmag:8.9, ra: '2h 42.7m', dec: '+00° 02′', constellation: 'Cet', colloquial_name: '');
-AstronomicObject.create(name: 'M78', equivalence: 'NGC2068', type_object: 'Di', bmag:8.3, ra: '5h 46.7m', dec: '+00° 03′', constellation: 'Ori', colloquial_name: '');
-AstronomicObject.create(name: 'M79', equivalence: 'NGC1904', type_object: 'Gc', bmag:7.7, ra: '5h 24.5m', dec: '-24° 33′', constellation: 'Lep', colloquial_name: '');
-AstronomicObject.create(name: 'M80', equivalence: 'NGC6093', type_object: 'Gc', bmag:7.3, ra: '16h 17m', dec: '-22° 59′', constellation: 'Sco', colloquial_name: '');
-AstronomicObject.create(name: 'M81', equivalence: 'NGC3031', type_object: 'Sp', bmag:6.9, ra: '9h 55.6m', dec: '+69° 04′', constellation: 'UMa', colloquial_name: 'Bodes Galaxy');
-AstronomicObject.create(name: 'M82', equivalence: 'NGC3034', type_object: 'Ir', bmag:8.4, ra: '9h 55.8m', dec: '+69° 41′', constellation: 'UMa', colloquial_name: 'Cigar Galaxy');
-AstronomicObject.create(name: 'M83', equivalence: 'NGC5236', type_object: 'Sp', bmag:7.6, ra: '13h 37m', dec: '-29° 52′', constellation: 'Hya', colloquial_name: 'Southern Pinwheel');
-AstronomicObject.create(name: 'M84', equivalence: 'NGC4374', type_object: 'Ln', bmag:9.1, ra: '12h 25.1m', dec: '+12° 53′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M85', equivalence: 'NGC4382', type_object: 'Ln', bmag:9.1, ra: '12h 25.5m', dec: '+18° 12′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M86', equivalence: 'NGC4406', type_object: 'Ln', bmag:8.9, ra: '12h 26.2m', dec: '+12° 57′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M87', equivalence: 'NGC4486', type_object: 'El', bmag:8.6, ra: '12h 30.8m', dec: '+12° 24′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M88', equivalence: 'NGC4501', type_object: 'Sp', bmag:9.6, ra: '12h 32.1m', dec: '+14° 26′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M89', equivalence: 'NGC4552', type_object: 'El', bmag:9.8, ra: '12h 35.7m', dec: '+12° 33′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M90', equivalence: 'NGC4569', type_object: 'Sp', bmag:9.5, ra: '12h 36.8m', dec: '+13° 10′', constellation: 'Vir', colloquial_name: '');
-AstronomicObject.create(name: 'M91', equivalence: 'NGC4548', type_object: 'Ba', bmag:10.2, ra: '12h 35.5m', dec: '+14° 30′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M92', equivalence: 'NGC6341', type_object: 'Gc', bmag:6.4, ra: '17h 17.1m', dec: '+43° 08′', constellation: 'Her', colloquial_name: '');
-AstronomicObject.create(name: 'M93', equivalence: 'NGC2447', type_object: 'Oc', bmag:6, ra: '7h 44.6m', dec: '-23° 52′', constellation: 'Pup', colloquial_name: '');
-AstronomicObject.create(name: 'M94', equivalence: 'NGC4736', type_object: 'Sp', bmag:8.2, ra: '12h 50.9m', dec: '+41° 08′', constellation: 'CVn', colloquial_name: '');
-AstronomicObject.create(name: 'M95', equivalence: 'NGC3351', type_object: 'Ba', bmag:9.7, ra: '10h 44m', dec: '+11° 42′', constellation: 'Leo', colloquial_name: '');
-AstronomicObject.create(name: 'M96', equivalence: 'NGC3368', type_object: 'Sp', bmag:9.2, ra: '10h 46.8m', dec: '+11° 49′', constellation: 'Leo', colloquial_name: '');
-AstronomicObject.create(name: 'M97', equivalence: 'NGC3587', type_object: 'Pl', bmag:9.9, ra: '11h 14.8m', dec: '+55° 01′', constellation: 'UMa', colloquial_name: 'Owl Nebula');
-AstronomicObject.create(name: 'M98', equivalence: 'NGC4192', type_object: 'Sp', bmag:10.1, ra: '12h 13.9m', dec: '+14° 55′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M99', equivalence: 'NGC4254', type_object: 'Sp', bmag:9.9, ra: '12h 18.9m', dec: '+14° 26′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M100', equivalence: 'NGC4321', type_object: 'Sp', bmag:9.3, ra: '12h 23m', dec: '+15° 50′', constellation: 'Com', colloquial_name: '');
-AstronomicObject.create(name: 'M101', equivalence: 'NGC5457', type_object: 'Sp', bmag:7.9, ra: '14h 03.2m', dec: '+54° 21′', constellation: 'UMa', colloquial_name: 'Pinwheel Galaxy');
-AstronomicObject.create(name: 'M102', equivalence: 'NGC5866', type_object: 'Ln', bmag:9.9, ra: '15h 06.5m', dec: '+55° 46′', constellation: 'Dra', colloquial_name: '');
-AstronomicObject.create(name: 'M103', equivalence: 'NGC581', type_object: 'Oc', bmag:7.4, ra: '1h 33.2m', dec: '+60° 42′', constellation: 'Cas', colloquial_name: '');
-AstronomicObject.create(name: 'M104', equivalence: 'NGC4594', type_object: 'Sp', bmag:8, ra: '12h 40m', dec: '-11° 37′', constellation: 'Vir', colloquial_name: 'Sombrero Galaxy');
-AstronomicObject.create(name: 'M105', equivalence: 'NGC3379', type_object: 'El', bmag:9.3, ra: '10h 47.8m', dec: '+12° 35′', constellation: 'Leo', colloquial_name: '');
-AstronomicObject.create(name: 'M106', equivalence: 'NGC4258', type_object: 'Sp', bmag:8.4, ra: '12h 18.9m', dec: '+47° 19′', constellation: 'CVn', colloquial_name: '');
-AstronomicObject.create(name: 'M107', equivalence: 'NGC6171', type_object: 'Gc', bmag:7.9, ra: '16h 32.5m', dec: '-13° 03′', constellation: 'Oph', colloquial_name: '');
-AstronomicObject.create(name: 'M108', equivalence: 'NGC3556', type_object: 'Sp', bmag:10, ra: '11h 11.5m', dec: '+55° 40′', constellation: 'UMa', colloquial_name: '');
-AstronomicObject.create(name: 'M109', equivalence: 'NGC3992', type_object: 'Ba', bmag:9.8, ra: '11h 57.6m', dec: '+53° 23′', constellation: 'UMa', colloquial_name: '');
-AstronomicObject.create(name: 'M110', equivalence: 'NGC205', type_object: 'El', bmag:8.5, ra: '00h 40.4m', dec: '+41° 41′', constellation: 'And', colloquial_name: '');
+# AstronomicObject.create(name: 'M1', equivalence: 'NGC1952', type_object: 'Sn', bmag:8.4, ra: '5h 34.5m', dec: '+22° 01′', constellation: 'Tau', colloquial_name: 'Crab Nebula');
+# AstronomicObject.create(name: 'M2', equivalence: 'NGC7089', type_object: 'Gc', bmag:6.5, ra: '21h 33.5m', dec: '+00° 49′', constellation: 'Aqr', colloquial_name: '');
+# AstronomicObject.create(name: 'M3', equivalence: 'NGC5272', type_object: 'Gc', bmag:6.2, ra: '13h 42.2m', dec: '+28° 23′', constellation: 'CVn', colloquial_name: '');
+# AstronomicObject.create(name: 'M4', equivalence: 'NGC6121', type_object: 'Gc', bmag:5.6, ra: '16h 23.6m', dec: '-26° 32′', constellation: 'Sco', colloquial_name: '');
+# AstronomicObject.create(name: 'M5', equivalence: 'NGC5904', type_object: 'Gc', bmag:5.6, ra: '15h 18.6m', dec: '+02° 05′', constellation: 'Ser', colloquial_name: '');
+# AstronomicObject.create(name: 'M6', equivalence: 'NGC6405', type_object: 'Oc', bmag:4.2, ra: '17h 40.1m', dec: '-32° 13′', constellation: 'Sco', colloquial_name: 'Butterfly Cluster');
+# AstronomicObject.create(name: 'M7', equivalence: 'NGC6475', type_object: 'Oc', bmag:3.3, ra: '17h 53.9m', dec: '-34° 49′', constellation: 'Sco', colloquial_name: 'Ptolemys Cluster');
+# AstronomicObject.create(name: 'M8', equivalence: 'NGC6523', type_object: 'Di', bmag:6, ra: '18h 03.8m', dec: '-24° 23′', constellation: 'Sgr', colloquial_name: 'Lagoon Nebula');
+# AstronomicObject.create(name: 'M9', equivalence: 'NGC6333', type_object: 'Gc', bmag:7.7, ra: '17h 19.2m', dec: '-18° 31′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M10', equivalence: 'NGC6254', type_object: 'Gc', bmag:6.6, ra: '16h 57.1m', dec: '-04° 06′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M11', equivalence: 'NGC6705', type_object: 'Oc', bmag:6.3, ra: '18h 51.1m', dec: '-06° 16′', constellation: 'Sct', colloquial_name: 'Wild Duck Cluster');
+# AstronomicObject.create(name: 'M12', equivalence: 'NGC6218', type_object: 'Gc', bmag:6.7, ra: '16h 47.2m', dec: '-01° 57′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M13', equivalence: 'NGC6205', type_object: 'Gc', bmag:5.8, ra: '16h 41.7m', dec: '+36° 28′', constellation: 'Her', colloquial_name: 'Great Hercules Globular');
+# AstronomicObject.create(name: 'M14', equivalence: 'NGC6402', type_object: 'Gc', bmag:7.6, ra: '17h 37.6m', dec: '-03° 15′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M15', equivalence: 'NGC7078', type_object: 'Gc', bmag:6.2, ra: '21h 30m', dec: '+12° 10′', constellation: 'Peg', colloquial_name: 'Great Pegasus Globular');
+# AstronomicObject.create(name: 'M16', equivalence: 'NGC6611', type_object: 'Oc', bmag:6.4, ra: '18h 18.8m', dec: '-13° 47′', constellation: 'Ser', colloquial_name: 'Eagle Nebula');
+# AstronomicObject.create(name: 'M17', equivalence: 'NGC6618', type_object: 'Di', bmag:7, ra: '18h 20.8m', dec: '-16° 11′', constellation: 'Sgr', colloquial_name: 'Omega Nebula');
+# AstronomicObject.create(name: 'M18', equivalence: 'NGC6613', type_object: 'Oc', bmag:7.5, ra: '18h 19.9m', dec: '-17° 08′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M19', equivalence: 'NGC6273', type_object: 'Gc', bmag:6.8, ra: '17h 02.6m', dec: '-26° 16′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M20', equivalence: 'NGC6514', type_object: 'Di', bmag:9, ra: '18h 02.6m', dec: '-23° 02′', constellation: 'Sgr', colloquial_name: 'Trifid Nebula');
+# AstronomicObject.create(name: 'M21', equivalence: 'NGC6531', type_object: 'Oc', bmag:6.5, ra: '18h 04.6m', dec: '-22° 30′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M22', equivalence: 'NGC6656', type_object: 'Gc', bmag:5.1, ra: '18h 36.4m', dec: '-23° 54′', constellation: 'Sgr', colloquial_name: 'Sagittarius Cluster');
+# AstronomicObject.create(name: 'M23', equivalence: 'NGC6494', type_object: 'Oc', bmag:6.9, ra: '17h 56.8m', dec: '-19° 01′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M24', equivalence: '', type_object: 'MW', bmag:4.6, ra: '18h 16.9m', dec: '-18° 30′', constellation: 'Sgr', colloquial_name: 'Sagittarius Star Cloud');
+# AstronomicObject.create(name: 'M25', equivalence: 'IC4725', type_object: 'Oc', bmag:6.5, ra: '18h 31.6m', dec: '-19° 15′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M26', equivalence: 'NGC6694', type_object: 'Oc', bmag:8, ra: '18h 45.2m', dec: '-09° 24′', constellation: 'Sct', colloquial_name: '');
+# AstronomicObject.create(name: 'M27', equivalence: 'NGC6853', type_object: 'Pl', bmag:7.4, ra: '19h 59.6m', dec: '+22° 43′', constellation: 'Vul', colloquial_name: 'Dumbbell Nebula');
+# AstronomicObject.create(name: 'M28', equivalence: 'NGC6626', type_object: 'Gc', bmag:6.8, ra: '18h 24.5m', dec: '-24° 52′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M29', equivalence: 'NGC6913', type_object: 'Oc', bmag:7.1, ra: '20h 23.9m', dec: '+38° 32′', constellation: 'Cyg', colloquial_name: '');
+# AstronomicObject.create(name: 'M30', equivalence: 'NGC7099', type_object: 'Gc', bmag:7.2, ra: '21h 40.4m', dec: '-23° 11′', constellation: 'Cap', colloquial_name: '');
+# AstronomicObject.create(name: 'M31', equivalence: 'NGC224', type_object: 'Sp', bmag:3.4, ra: '0h 41.8m', dec: '+41° 16′', constellation: 'And', colloquial_name: 'Andromeda Galaxy');
+# AstronomicObject.create(name: 'M32', equivalence: 'NGC221', type_object: 'El', bmag:8.1, ra: '0h 42.8m', dec: '+40° 52′', constellation: 'And', colloquial_name: '');
+# AstronomicObject.create(name: 'M33', equivalence: 'NGC598', type_object: 'Sp', bmag:5.7, ra: '1h 33.9m', dec: '+30° 39′', constellation: 'Tri', colloquial_name: 'Triangulum Galaxy');
+# AstronomicObject.create(name: 'M34', equivalence: 'NGC1039', type_object: 'Oc', bmag:5.5, ra: '2h 42m', dec: '+42° 47′', constellation: 'Per', colloquial_name: '');
+# AstronomicObject.create(name: 'M35', equivalence: 'NGC2168', type_object: 'Oc', bmag:5.3, ra: '6h 08.9m', dec: '+24° 20′', constellation: 'Gem', colloquial_name: '');
+# AstronomicObject.create(name: 'M36', equivalence: 'NGC1960', type_object: 'Oc', bmag:6.3, ra: '5h 36.1m', dec: '+34° 08′', constellation: 'Aur', colloquial_name: '');
+# AstronomicObject.create(name: 'M37', equivalence: 'NGC2099', type_object: 'Oc', bmag:6.2, ra: '5h 52.4m', dec: '+32° 33′', constellation: 'Aur', colloquial_name: '');
+# AstronomicObject.create(name: 'M38', equivalence: 'NGC1912', type_object: 'Oc', bmag:7.4, ra: '5h 28.7m', dec: '+35° 50′', constellation: 'Aur', colloquial_name: '');
+# AstronomicObject.create(name: 'M39', equivalence: 'NGC7092', type_object: 'Oc', bmag:4.6, ra: '21h 32.2m', dec: '+48° 26′', constellation: 'Cyg', colloquial_name: '');
+# AstronomicObject.create(name: 'M40', equivalence: '', type_object: 'Ds', bmag:8.4, ra: '12h 22.4m', dec: '+58° 05′', constellation: 'UMa', colloquial_name: 'Winnecke 4');
+# AstronomicObject.create(name: 'M41', equivalence: 'NGC2287', type_object: 'Oc', bmag:4.6, ra: '6h 47m', dec: '-20° 44′', constellation: 'CMa', colloquial_name: '');
+# AstronomicObject.create(name: 'M42', equivalence: 'NGC1976', type_object: 'Di', bmag:4, ra: '5h 35.4m', dec: '-05° 27′', constellation: 'Ori', colloquial_name: 'Great Nebula in Orion');
+# AstronomicObject.create(name: 'M43', equivalence: 'NGC1982', type_object: 'Di', bmag:9, ra: '5h 35.6m', dec: '-05° 16′', constellation: 'Ori', colloquial_name: 'De Mairans Nebula');
+# AstronomicObject.create(name: 'M44', equivalence: 'NGC2632', type_object: 'Oc', bmag:3.7, ra: '8h 40.1m', dec: '+19° 59′', constellation: 'Cnc', colloquial_name: 'Beehive Cluster');
+# AstronomicObject.create(name: 'M45', equivalence: '', type_object: 'Oc', bmag:1.6, ra: '3h 47m', dec: '+24° 07′', constellation: 'Tau', colloquial_name: 'Pleiades');
+# AstronomicObject.create(name: 'M46', equivalence: 'NGC2437', type_object: 'Oc', bmag:6, ra: '7h 41.8m', dec: '-14° 49′', constellation: 'Pup', colloquial_name: '');
+# AstronomicObject.create(name: 'M47', equivalence: 'NGC2422', type_object: 'Oc', bmag:5.2, ra: '7h 36.6m', dec: '-14° 30′', constellation: 'Pup', colloquial_name: '');
+# AstronomicObject.create(name: 'M48', equivalence: 'NGC2548', type_object: 'Oc', bmag:5.5, ra: '8h 13.8m', dec: '-05° 48′', constellation: 'Hya', colloquial_name: '');
+# AstronomicObject.create(name: 'M49', equivalence: 'NGC4472', type_object: 'El', bmag:8.4, ra: '12h 29.8m', dec: '+08° 00′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M50', equivalence: 'NGC2323', type_object: 'Oc', bmag:6.3, ra: '7h 03.2m', dec: '-08° 20′', constellation: 'Mon', colloquial_name: '');
+# AstronomicObject.create(name: 'M51', equivalence: 'NGC5194', type_object: 'Sp', bmag:8.4, ra: '13h 30m', dec: '+47° 11′', constellation: 'CVn', colloquial_name: 'Whirlpool Galaxy');
+# AstronomicObject.create(name: 'M52', equivalence: 'NGC7654', type_object: 'Oc', bmag:7.3, ra: '23h 24.2m', dec: '+61° 35′', constellation: 'Cas', colloquial_name: '');
+# AstronomicObject.create(name: 'M53', equivalence: 'NGC5024', type_object: 'Gc', bmag:7.6, ra: '13h 12.9m', dec: '+18° 10′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M54', equivalence: 'NGC6715', type_object: 'Gc', bmag:7.6, ra: '18h 55.1m', dec: '-30° 29′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M55', equivalence: 'NGC6809', type_object: 'Gc', bmag:6.3, ra: '19h 40m', dec: '-30° 58′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M56', equivalence: 'NGC6779', type_object: 'Gc', bmag:8.3, ra: '19h 16.6m', dec: '+30° 11′', constellation: 'Lyr', colloquial_name: '');
+# AstronomicObject.create(name: 'M57', equivalence: 'NGC6720', type_object: 'Pl', bmag:8.8, ra: '18h 53.6m', dec: '+33° 02′', constellation: 'Lyr', colloquial_name: 'Ring Nebula');
+# AstronomicObject.create(name: 'M58', equivalence: 'NGC4579', type_object: 'Ba', bmag:9.7, ra: '12h 37.7m', dec: '+11° 49′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M59', equivalence: 'NGC4621', type_object: 'El', bmag:9.6, ra: '12h 42m', dec: '+11° 39′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M60', equivalence: 'NGC4649', type_object: 'El', bmag:8.8, ra: '12h 43.7m', dec: '+11° 33′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M61', equivalence: 'NGC4303', type_object: 'Sp', bmag:9.7, ra: '12h 21.9m', dec: '+04° 28′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M62', equivalence: 'NGC6266', type_object: 'Gc', bmag:6.5, ra: '17h 01.2m', dec: '-30° 07′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M63', equivalence: 'NGC5055', type_object: 'Sp', bmag:8.6, ra: '13h 15.8m', dec: '+42° 02′', constellation: 'CVn', colloquial_name: 'Sunflower Galaxy');
+# AstronomicObject.create(name: 'M64', equivalence: 'NGC4826', type_object: 'Sp', bmag:8.5, ra: '12h 56.7m', dec: '+21° 41′', constellation: 'Com', colloquial_name: 'Black Eye Galaxy');
+# AstronomicObject.create(name: 'M65', equivalence: 'NGC3623', type_object: 'Sp', bmag:9.3, ra: '11h 18.9m', dec: '+13° 05′', constellation: 'Leo', colloquial_name: '');
+# AstronomicObject.create(name: 'M66', equivalence: 'NGC3627', type_object: 'Sp', bmag:8.9, ra: '11h 20.2m', dec: '+12° 59′', constellation: 'Leo', colloquial_name: '');
+# AstronomicObject.create(name: 'M67', equivalence: 'NGC2682', type_object: 'Oc', bmag:6.1, ra: '8h 50.4m', dec: '+11° 49′', constellation: 'Cnc', colloquial_name: '');
+# AstronomicObject.create(name: 'M68', equivalence: 'NGC4590', type_object: 'Gc', bmag:7.8, ra: '12h 39.5m', dec: '-26° 45′', constellation: 'Hya', colloquial_name: '');
+# AstronomicObject.create(name: 'M69', equivalence: 'NGC6637', type_object: 'Gc', bmag:7.6, ra: '18h 31.4m', dec: '-32° 21′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M70', equivalence: 'NGC6681', type_object: 'Gc', bmag:7.9, ra: '18h 43.2m', dec: '-32° 18′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M71', equivalence: 'NGC6838', type_object: 'Gc', bmag:8.2, ra: '19h 53.8m', dec: '+18° 47′', constellation: 'Sge', colloquial_name: '');
+# AstronomicObject.create(name: 'M72', equivalence: 'NGC6981', type_object: 'Gc', bmag:9.3, ra: '20h 53.5m', dec: '-12° 32′', constellation: 'Aqr', colloquial_name: '');
+# AstronomicObject.create(name: 'M73', equivalence: 'NGC6994', type_object: 'As', bmag:9, ra: '20h 59m', dec: '-12° 38′', constellation: 'Aqr', colloquial_name: '');
+# AstronomicObject.create(name: 'M74', equivalence: 'NGC628', type_object: 'Sp', bmag:9.4, ra: '1h 36.7m', dec: '+15° 47′', constellation: 'Psc', colloquial_name: '');
+# AstronomicObject.create(name: 'M75', equivalence: 'NGC6864', type_object: 'Gc', bmag:8.5, ra: '20h 06.1m', dec: '-21° 55′', constellation: 'Sgr', colloquial_name: '');
+# AstronomicObject.create(name: 'M76', equivalence: 'NGC650', type_object: 'Pl', bmag:10.1, ra: '1h 42.4m', dec: '+51° 34′', constellation: 'Per', colloquial_name: 'Little Dumbbell Nebula');
+# AstronomicObject.create(name: 'M77', equivalence: 'NGC1068', type_object: 'Sp', bmag:8.9, ra: '2h 42.7m', dec: '+00° 02′', constellation: 'Cet', colloquial_name: '');
+# AstronomicObject.create(name: 'M78', equivalence: 'NGC2068', type_object: 'Di', bmag:8.3, ra: '5h 46.7m', dec: '+00° 03′', constellation: 'Ori', colloquial_name: '');
+# AstronomicObject.create(name: 'M79', equivalence: 'NGC1904', type_object: 'Gc', bmag:7.7, ra: '5h 24.5m', dec: '-24° 33′', constellation: 'Lep', colloquial_name: '');
+# AstronomicObject.create(name: 'M80', equivalence: 'NGC6093', type_object: 'Gc', bmag:7.3, ra: '16h 17m', dec: '-22° 59′', constellation: 'Sco', colloquial_name: '');
+# AstronomicObject.create(name: 'M81', equivalence: 'NGC3031', type_object: 'Sp', bmag:6.9, ra: '9h 55.6m', dec: '+69° 04′', constellation: 'UMa', colloquial_name: 'Bodes Galaxy');
+# AstronomicObject.create(name: 'M82', equivalence: 'NGC3034', type_object: 'Ir', bmag:8.4, ra: '9h 55.8m', dec: '+69° 41′', constellation: 'UMa', colloquial_name: 'Cigar Galaxy');
+# AstronomicObject.create(name: 'M83', equivalence: 'NGC5236', type_object: 'Sp', bmag:7.6, ra: '13h 37m', dec: '-29° 52′', constellation: 'Hya', colloquial_name: 'Southern Pinwheel');
+# AstronomicObject.create(name: 'M84', equivalence: 'NGC4374', type_object: 'Ln', bmag:9.1, ra: '12h 25.1m', dec: '+12° 53′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M85', equivalence: 'NGC4382', type_object: 'Ln', bmag:9.1, ra: '12h 25.5m', dec: '+18° 12′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M86', equivalence: 'NGC4406', type_object: 'Ln', bmag:8.9, ra: '12h 26.2m', dec: '+12° 57′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M87', equivalence: 'NGC4486', type_object: 'El', bmag:8.6, ra: '12h 30.8m', dec: '+12° 24′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M88', equivalence: 'NGC4501', type_object: 'Sp', bmag:9.6, ra: '12h 32.1m', dec: '+14° 26′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M89', equivalence: 'NGC4552', type_object: 'El', bmag:9.8, ra: '12h 35.7m', dec: '+12° 33′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M90', equivalence: 'NGC4569', type_object: 'Sp', bmag:9.5, ra: '12h 36.8m', dec: '+13° 10′', constellation: 'Vir', colloquial_name: '');
+# AstronomicObject.create(name: 'M91', equivalence: 'NGC4548', type_object: 'Ba', bmag:10.2, ra: '12h 35.5m', dec: '+14° 30′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M92', equivalence: 'NGC6341', type_object: 'Gc', bmag:6.4, ra: '17h 17.1m', dec: '+43° 08′', constellation: 'Her', colloquial_name: '');
+# AstronomicObject.create(name: 'M93', equivalence: 'NGC2447', type_object: 'Oc', bmag:6, ra: '7h 44.6m', dec: '-23° 52′', constellation: 'Pup', colloquial_name: '');
+# AstronomicObject.create(name: 'M94', equivalence: 'NGC4736', type_object: 'Sp', bmag:8.2, ra: '12h 50.9m', dec: '+41° 08′', constellation: 'CVn', colloquial_name: '');
+# AstronomicObject.create(name: 'M95', equivalence: 'NGC3351', type_object: 'Ba', bmag:9.7, ra: '10h 44m', dec: '+11° 42′', constellation: 'Leo', colloquial_name: '');
+# AstronomicObject.create(name: 'M96', equivalence: 'NGC3368', type_object: 'Sp', bmag:9.2, ra: '10h 46.8m', dec: '+11° 49′', constellation: 'Leo', colloquial_name: '');
+# AstronomicObject.create(name: 'M97', equivalence: 'NGC3587', type_object: 'Pl', bmag:9.9, ra: '11h 14.8m', dec: '+55° 01′', constellation: 'UMa', colloquial_name: 'Owl Nebula');
+# AstronomicObject.create(name: 'M98', equivalence: 'NGC4192', type_object: 'Sp', bmag:10.1, ra: '12h 13.9m', dec: '+14° 55′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M99', equivalence: 'NGC4254', type_object: 'Sp', bmag:9.9, ra: '12h 18.9m', dec: '+14° 26′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M100', equivalence: 'NGC4321', type_object: 'Sp', bmag:9.3, ra: '12h 23m', dec: '+15° 50′', constellation: 'Com', colloquial_name: '');
+# AstronomicObject.create(name: 'M101', equivalence: 'NGC5457', type_object: 'Sp', bmag:7.9, ra: '14h 03.2m', dec: '+54° 21′', constellation: 'UMa', colloquial_name: 'Pinwheel Galaxy');
+# AstronomicObject.create(name: 'M102', equivalence: 'NGC5866', type_object: 'Ln', bmag:9.9, ra: '15h 06.5m', dec: '+55° 46′', constellation: 'Dra', colloquial_name: '');
+# AstronomicObject.create(name: 'M103', equivalence: 'NGC581', type_object: 'Oc', bmag:7.4, ra: '1h 33.2m', dec: '+60° 42′', constellation: 'Cas', colloquial_name: '');
+# AstronomicObject.create(name: 'M104', equivalence: 'NGC4594', type_object: 'Sp', bmag:8, ra: '12h 40m', dec: '-11° 37′', constellation: 'Vir', colloquial_name: 'Sombrero Galaxy');
+# AstronomicObject.create(name: 'M105', equivalence: 'NGC3379', type_object: 'El', bmag:9.3, ra: '10h 47.8m', dec: '+12° 35′', constellation: 'Leo', colloquial_name: '');
+# AstronomicObject.create(name: 'M106', equivalence: 'NGC4258', type_object: 'Sp', bmag:8.4, ra: '12h 18.9m', dec: '+47° 19′', constellation: 'CVn', colloquial_name: '');
+# AstronomicObject.create(name: 'M107', equivalence: 'NGC6171', type_object: 'Gc', bmag:7.9, ra: '16h 32.5m', dec: '-13° 03′', constellation: 'Oph', colloquial_name: '');
+# AstronomicObject.create(name: 'M108', equivalence: 'NGC3556', type_object: 'Sp', bmag:10, ra: '11h 11.5m', dec: '+55° 40′', constellation: 'UMa', colloquial_name: '');
+# AstronomicObject.create(name: 'M109', equivalence: 'NGC3992', type_object: 'Ba', bmag:9.8, ra: '11h 57.6m', dec: '+53° 23′', constellation: 'UMa', colloquial_name: '');
+# AstronomicObject.create(name: 'M110', equivalence: 'NGC205', type_object: 'El', bmag:8.5, ra: '00h 40.4m', dec: '+41° 41′', constellation: 'And', colloquial_name: '');
 
-AstronomicObject.create(name: 'Mercury', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Venus', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Mars', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Jupiter', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Saturn', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Uranus', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Neptune', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Pluto', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Moon', catalog: 'SolarSistem', type_object: 'Moon',  constellation: 'Solar Sistem', colloquial_name: '');
-AstronomicObject.create(name: 'Sun', catalog: 'SolarSistem', type_object: 'Star',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Mercury', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Venus', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Mars', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Jupiter', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Saturn', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Uranus', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Neptune', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Pluto', catalog: 'SolarSistem', type_object: 'Planet',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Moon', catalog: 'SolarSistem', type_object: 'Moon',  constellation: 'Solar Sistem', colloquial_name: '');
+# AstronomicObject.create(name: 'Sun', catalog: 'SolarSistem', type_object: 'Star',  constellation: 'Solar Sistem', colloquial_name: '');
+
+
+TypeObject.create(type_name: 'Oc', name:  'Open Cluster');
+TypeObject.create(type_name: 'Sp', name:  'Spiral Galaxy');
+TypeObject.create(type_name: 'Gc', name:  'Globular Cluster');    
+TypeObject.create(type_name: 'Ba', name:  'Barred Galaxy');
+TypeObject.create(type_name: 'Pl', name:  'Planetary Nebula');    
+TypeObject.create(type_name: 'Ln', name:  'Lenticular Galaxy');
+TypeObject.create(type_name: 'Di', name:  'Diffuse Nebula');
+TypeObject.create(type_name: 'El', name:  'Elliptical Galaxy');
+TypeObject.create(type_name: 'As', name:  'Asterism');    
+TypeObject.create(type_name: 'Ir', name:  'Irregular Galaxy');
+TypeObject.create(type_name: 'Ds', name:  'Double Star');   
+TypeObject.create(type_name: 'Sn', name:  'Supernova Remnant');
+TypeObject.create(type_name: 'MW', name:  'Milky Way Patch');
+TypeObject.create(type_name: 'Star', name:  'Star');
+TypeObject.create(type_name: 'Planet', name:  'Planet');
+TypeObject.create(type_name: 'Moon', name:  'Moon');
+
+
+
